Addresses the sql alchemy issue where newer versions renamed exception class
The real fix is to be okay with getting the latest sql alchemy and
changing where we import the exception class.  This solution was
reported a few months back by Prashanth.  In the mean time we are
simply capping the version of sql alchemy such that we don't run into
this.

The error manifests as:
(this particular example is the node manager project, but the same
issue is with security and dashboard - which also use sql alchemy)

Traceback (most recent call last):
  File "/usr/local/uvcdat/bin/esgf_node_manager_initialize", line 5, in <module>
    pkg_resources.run_script('esgf-node-manager==0.1.5', 'esgf_node_manager_initialize')
  File "build/bdist.linux-x86_64/egg/pkg_resources.py", line 489, in run_script
  File "build/bdist.linux-x86_64/egg/pkg_resources.py", line 1207, in run_script
  File "/usr/local/uvcdat/lib/python2.7/site-packages/esgf_node_manager-0.1.5-py2.7.egg/EGG-INFO/scripts/esgf_node_manager_initialize", line 10, in <module>
    from migrate.versioning.exceptions import DatabaseAlreadyControlledError
ImportError: No module named exceptions
ERROR: Could not create esgf manager database tables in esgcet